Use Files::copy from terracotta utilities
